create base cube
Create a common base class for standard and transparent cubes
