+++ src/polycubed/src/service_controller.h
@@ -65,16 +65,16 @@ class ServiceController {
   std::string get_swagger_codegen_git_repo_id() const;
   std::string get_servicecontroller() const;
   std::string get_datamodel() const;
-  std::vector<std::shared_ptr<CubeIface>> get_cubes();
+  std::vector<std::shared_ptr<BaseCubeIface>> get_cubes();
 
   // Instantiate the managementGrpc Object, using the endpoint url
   void connect(std::string PolycubeEndpoint);
   ServiceControllerType get_type() const;
 
-  static std::shared_ptr<CubeIface> get_cube(const std::string &name);
-  static std::vector<std::shared_ptr<CubeIface>> get_all_cubes();
+  static std::shared_ptr<BaseCubeIface> get_cube(const std::string &name);
+  static std::vector<std::shared_ptr<BaseCubeIface>> get_all_cubes();
 
-  static void register_cube(std::shared_ptr<CubeIface> cube,
+  static void register_cube(std::shared_ptr<BaseCubeIface> cube,
                             const std::string &service);
   static void unregister_cube(const std::string &name);
 
@@ -98,8 +98,8 @@ class ServiceController {
                               std::string &port);
 
   // these objects save all the common objects accross different services
-  static std::unordered_map<std::string, std::shared_ptr<CubeIface>> cubes;
   static std::unordered_map<Guid, std::unique_ptr<Node>> ports_to_ifaces;
+  static std::unordered_map<std::string, std::shared_ptr<BaseCubeIface>> cubes;
   static std::unordered_map<std::string, std::string> ports_to_ports;
 
   static std::unordered_map<std::string, std::string> cubes_x_service;
